Further changes per review comments.
